Implement reconnect function and service in Dyson PureHotCool
The approach is hacky and just copied the way it was done for the Dyson
PureCool. I know there is a better way, but since I don't know it and
this will work, I will leave it as it is.
Also, [skip-ci] because I haven't migrated the CircleCI workflow to use
docker-compose and the new custom built docker image.