Add notification when alarm is triggered pending
Set the URL for the alarm sound in the file instead of in a secret
because Home Assistant does not play the media file when stored in the
secrets file. This will make it necessary to update the configuration
when changing the host. Can't use a domain name because Google speakers
do not resolve local network names.